% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_correction.R
\name{compute_index1and2_pixel}
\alias{compute_index1and2_pixel}
\title{Title}
\usage{
compute_index1and2_pixel(log_temp, search_area, secs_per_temp, delta)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
