% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_correction.R
\name{compute_index3_pixel}
\alias{compute_index3_pixel}
\title{Title}
\usage{
compute_index3_pixel(
  posn,
  center_x,
  center_y,
  log_temp_u,
  log_temp_pro,
  num_sample,
  sample_size,
  search_width,
  search_height,
  nrows,
  ncols,
  lat_step,
  long_step
)
}
\arguments{
\item{long_step}{}
}
\description{
Title
}
