% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_correction.R
\name{compute_SST_lat}
\alias{compute_SST_lat}
\title{Title}
\usage{
compute_SST_lat(
  filename,
  log_posn,
  log_temp,
  n_limit,
  s_limit,
  search_box_width,
  search_box_height,
  delta_temp,
  secs_per_temp,
  delta_lat,
  start_lat,
  start_long
)
}
\arguments{
\item{start_long}{}
}
\description{
Title
}
