% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_GLSHelper_folders}
\alias{create_GLSHelper_folders}
\title{Create GLShelper folder structure}
\usage{
create_GLSHelper_folders(root = NULL)
}
\arguments{
\item{root}{[character]\cr Required. The root folder of your data analysis
project.}
}
\value{
Nothing.
}
\description{
blah
}
\details{
GLSHelper is obnoxiously opinionated about the folder structure
containing your logger data and any generated outputs, shapefiles, etc.

\if{html}{\out{<div class="sourceCode">}}\preformatted{-- root
  |__Data
    |__geolocation settings.csv
    |__Tag data
       |__MK3005 025
         ...
  |__GIS
     |__Shapefiles
     |__Images
}\if{html}{\out{</div>}}

This function checks to see if sub-folders \code{Data, Data\\Tag Data, GIS,
  GIS\\Shapefiles} and \code{GIS\\Images} exist and creates them if not.
}
\section{Author}{
 Dave Fifield
}

\examples{

# Create folder structure in root of RStudio project
create_GLSHelper_folders(here::here())
}
