% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_all_geolocation}
\alias{do_all_geolocation}
\title{Do geolocation for multiple GLS data sets}
\usage{
do_all_geolocation(folder, file)
}
\arguments{
\item{folder}{[character]\cr Required. The name of a folder which contains
the GLS light data for one or more devices. The light data for each device
must be in a separate sub-folder of \code{folder}. The name of each
sub-folder must match the \code{tagName} column of the config file. See
the \code{file} parameter.}

\item{file}{[character]\cr Required. The full pathname of a CSV configuration
file. This file contains one line per GLS data set to
be processed. See \code{\link{do_geolocation}} for details.}
}
\value{
Nothing.
}
\description{
Process multiple GLS logger raw light data sets according to
settings given in a configuration CSV file.
}
\details{
This function is a wrapper for \code{\link{do_geolocation()}}. It
reads the config file specified by \code{folder} and \code{file} and
calls \code{do.geolocation} repeatedly - once for each data set.

There are quite strict naming conventions for subfolders, light files, etc.
See \code{\link{create_GLSHelper_folders}} for details.
}
\section{Author}{
 Dave Fifield
}

