% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_geolocation}
\alias{do_geolocation}
\title{Do geolocation for one GLS data set}
\usage{
do_geolocation(cfg, folder, elev = NULL)
}
\arguments{
\item{cfg}{[character]\cr Required. A dataframe of configuration information
to control the processing.}

\item{folder}{[character]\cr Required. The path to the folder that contains
the tag data stored in separate sub-folders for each tag. The name of
each sub-folder must match the \code{cfg$tagName}.}

\item{elev}{[character]\cr Required. elevation angle to use, omit to use
elevation angle specified in \code{cfg}.}
}
\value{
Nothing.
}
\description{
Process a single GLS logger raw light data set according to
settings given in \code{cfg} using the \pkg{GeoLight} package.
}
\details{
The details

There are quite strict naming conventions for subfolders, light files, etc.
}
\section{Author}{
 Dave Fifield
}

