% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_multi_geolocation}
\alias{do_multi_geolocation}
\title{Do geolocation for multiple GLS data sets}
\usage{
do_multi_geolocation(folder, cfgfile, shapefolder = NULL, subset = NULL)
}
\arguments{
\item{folder}{[character]\cr Required. The name of a folder which contains
the GLS light data for one or more devices. The light data for each device
must be in a separate sub-folder of \code{folder}. The name of each
sub-folder must match the \code{tagName} column of the config file. See
the \code{file} parameter.}

\item{cfgfile}{[character]\cr Required. The full pathname of a CSV configuration
file. This file contains one line per GLS data set to
be processed. See \code{\link{do_geolocation}} for details.}

\item{shapefolder}{[character]\cr optional. The full pathname to a folder
where shapefiles will be stored (if requested),}

\item{subset}{[character]\cr Optional. A character vector of tag names to
include. These must match the \code{tagName} column in the configuration
file. This argument overrides the value of the \code{include} column
in the configuration file.}
}
\value{
A list which each element being the return value of
\code{\link{do_geolocation}}.
}
\description{
Process multiple GLS logger raw light data sets according to
settings given in a configuration CSV file.
}
\details{
This function
reads the config file specified by \code{cfgfile} - each row
specifies the settings for one dataset to process.  It then
calls \url{do.geolocation} repeatedly - once for each data set.

There are quite strict naming conventions for subfolders, light files, etc.
See \code{\link{create_GLSHelper_folders}} for details.
}
\section{Author}{
 Dave Fifield
}

