% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_correction.R
\name{do_sst_correction}
\alias{do_sst_correction}
\title{Do SST correction for a GLS dataset}
\usage{
do_sst_correction(options)
}
\arguments{
\item{options}{(required) List of options to control the SST correction...}
}
\description{
Do SST correction for a GLS dataset
}
