% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activity data}
\usage{
plot_activity(dat, maxact = 200)
}
\arguments{
\item{dat}{[dataframe]\cr Required. A dataframe containing columns
\code{datetime} and \code{act}.}

\item{maxact}{[numeric]\cr optional - default 200. The maximum value \code{act} can take.
Used to scale the y axis of the plot in \%.}
}
\value{
The \code{ggplot} invisibly.
}
\description{
Plot activity data as measured by wet/dry switch
}
\details{
This function takes the activity data from the wet-dry sensor and
plots time on the x-axis and percentage wet per interval on the y-axis.
The length of the interval is computed as the time between the first two
points in the activity series. This interval is simply used to label the
y-axis.
}
\section{Author}{
 Dave Fifield
}

