% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_correction.R
\name{read_GLS_posns}
\alias{read_GLS_posns}
\title{Read GLS positions}
\usage{
read_GLS_posns(options)
}
\arguments{
\item{options}{\link{required} List of options at least containing named element
"posnfile" giving the full pathname to the position file to be read. This
file must be either a \code{csv} or Excel (\code{.xls} or \code{.xlsx}) file.}
}
\value{
A tibble containing columns for \code(Date), \code{Longitude}, and
\code{Latitude} plus any other columns present in the file
}
\description{
Read GLS positions
}
